Replaces manual modulo checks with is_multiple_of and refines mode handling
Updates multiple locations to use the is_multiple_of method for clarity
and potential efficiency over manual modulo operations. Refines
permission and mode handling by removing unnecessary type casts,
improving code readability and correctness in permission assignments.

+++ fs/kfs/src/fs/ext4/fs.rs
@@ -83,7 +83,7 @@ impl FilesystemOps for Ext4Filesystem {
     fn stat(&self) -> VfsResult<StatFs> {
         let fs = self.lock();
         let superblock = &fs.fs.superblock;
-        let block_size = superblock.block_size() as u64;
+        let block_size = superblock.block_size();
         let blocks = superblock.blocks_count();
         let blocks_free = superblock.free_blocks_count();
         Ok(StatFs {
